More refactoring in shader and materials

+++ Engine/Source/Objects/Transform.h
@@ -39,9 +39,9 @@ namespace ActaEngine
 
 		glm::vec3 direction;
 
-		float yaw = 0.0f;
-		float pitch = 0.0f;
-		float roll = 0.0f;
+		double yaw = 0.0;
+		double pitch = 0.0;
+		double roll = 0.0;
 
 		float angles = 0.0f;
 	public:
